fixed: addCommentMutation, Comment dto, service

+++ src/main/java/com/example/RestTest/repository/CommentsRepository.java
@@ -3,5 +3,5 @@ package com.example.RestTest.repository;
 import com.example.RestTest.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentsRepository extends JpaRepository<Comment, Long> {
+public interface CommentsRepository extends JpaRepository<Comment, Integer> {
 }
